refactor(eve): isolate service output resolution
